server/requirements.txt
Flask==2.1.1
# Werkzeug==2.0.3
google-generativeai==0.5.0
pypdf==4.2.0
python-dotenv==0.20.0
werkzeug==2.3.7
fastapi
uvicorn